+++ src/main/java/com/pestphp/pest/configuration/PestRunConfigurationSettings.java
@@ -1,8 +1,0 @@
-package com.pestphp.pest.configuration;
-
-import com.jetbrains.php.testFramework.run.PhpTestRunConfigurationSettings;
-import org.jetbrains.annotations.NotNull;
-
-public class PestRunConfigurationSettings extends PhpTestRunConfigurationSettings {
-    @NotNull private final PestRunnerSettings pestRunnerSettings = new PestRunnerSettings();
-}