+++ src/main/java/com/pestphp/pest/configuration/PestRunnerSettings.java
@@ -1,4 +1,0 @@
-package com.pestphp.pest.configuration;
-
-public class PestRunnerSettings {
-}