+++ src/main/kotlin/com/pestphp/pest/PestBundle.kt
@@ -1,0 +1,14 @@
+package com.pestphp.pest
+
+import com.intellij.AbstractBundle
+import org.jetbrains.annotations.NonNls
+import org.jetbrains.annotations.PropertyKey
+
+@NonNls
+private const val BUNDLE = "pestBundle"
+
+object PestBundle : AbstractBundle(BUNDLE) {
+    @Suppress("SpreadOperator")
+    @JvmStatic
+    fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) = getMessage(key, *params)
+}