+++ src/main/kotlin/com/pestphp/pest/PestIcons.kt
@@ -1,0 +1,10 @@
+package com.pestphp.pest
+
+import com.intellij.openapi.util.IconLoader
+
+object PestIcons {
+    val LOGO = IconLoader.getIcon("/logo.svg")
+    val RUN_SINGLE_TEST = IconLoader.getIcon("/pest-run.svg")
+    val CONFIG = IconLoader.getIcon("/pest-config.svg")
+    val FILE = IconLoader.getIcon("/pest-file.svg")
+}