+++ src/main/kotlin/com/pestphp/pest/configuration/PestRunConfigurationSettings.kt
@@ -1,0 +1,7 @@
+package com.pestphp.pest.configuration
+
+import com.jetbrains.php.testFramework.run.PhpTestRunConfigurationSettings
+
+class PestRunConfigurationSettings : PhpTestRunConfigurationSettings() {
+    private val pestRunnerSettings = PestRunnerSettings()
+}