+++ src/main/kotlin/com/pestphp/pest/runner/LocationInfo.kt
@@ -1,0 +1,8 @@
+package com.pestphp.pest.runner
+
+import com.intellij.openapi.vfs.VirtualFile
+
+class LocationInfo(
+    val file: VirtualFile,
+    val testName: String?
+)