+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 - Add running support with debugger ([#19](https://github.com/pestphp/pest-intellij/pull/19))
 - `$this->field` type support for fields declared in `beforeEach` and `beforeAll` functions ([#22](https://github.com/pestphp/pest-intellij/pull/22))
 - Run tests scoped based on the cursor ([#24](https://github.com/pestphp/pest-intellij/pull/24), [#26](https://github.com/pestphp/pest-intellij/pull/26))
+- Added support for rerun tests when using new pest version ([#39](https://github.com/pestphp/pest-intellij/pull/39))
+- Added coverage support with clover output ([#39](https://github.com/pestphp/pest-intellij/pull/39))
+
+### Changed
+- Migrated all Java classes to Kotlin
 
 ### Fixed
 - Plugin require restart as PhpTestFrameworkType does not support dynamic plugins